ci: clean up env vars in actions, change dist paths and rollup config
