+++ .github/workflows/notion-pr-sync.yml
@@ -47,7 +47,7 @@ jobs:
   request:
     runs-on: ubuntu-latest
     steps:
-      - name: Sync Github PRs to Notion
+      - name: Sync GitHub PRs to Notion
         uses: sivashanmukh/github-notion-pr-sync@1.0.0
         with:
           notionKey: ${{ secrets.NOTION_BOT_KEY }}
