--- buildspec.prod.yaml
@@ -1,20 +1,0 @@
-version: 0.2
-
-phases:
-  install:
-    runtime-versions:
-      nodejs: 16
-  build:
-    commands:
-      - npm ci --unsafe-perm
-      - npm run test:ci
-      - npm run build:browser
-      - npm run build:integration:all
-      - sed -i -e 's|{{RS_BUGSNAG_API_KEY}}|'$RS_BUGSNAG_API_KEY'|' dist/rudder-analytics.min.js
-      - aws s3 cp dist/rudder-analytics.min.js s3://$S3_BUCKET_NAME/v1.1/rudder-analytics.min.js --cache-control max-age=3600 --acl public-read
-      - aws s3 cp dist/rudder-analytics.min.js.map s3://$S3_BUCKET_NAME/v1.1/rudder-analytics.min.js.map --cache-control max-age=3600 --acl public-read
-      - aws s3 cp dist/integrations/ s3://$S3_BUCKET_NAME/v1.1/js-integrations/ --recursive --cache-control max-age=3600 --acl public-read
-      - aws cloudfront create-invalidation --distribution-id $CLOUDFRONT_DISTRIBUTION_ID --paths "/v1.1/rudder-analytics.min.js" "/v1.1/rudder-analytics.min.js.map" "/v1.1/js-integrations*"
-artifacts:
-  files:
-    - "**/*"
