+++ package.json
@@ -5,16 +5,16 @@
   "main": "./dist/browser.min.js",
   "scripts": {
     "setup": "npm i",
-    "test": "npm run build:browser && sed 's/var rudderanalytics/rudderanalytics/g' dist/rudder-analytics.min.js > __tests__/prodsdk.js && jest --watchAll=false --maxWorkers=50% --coverage",
-    "test:ci": "npm run build:browser && sed 's/var rudderanalytics/rudderanalytics/g' dist/rudder-analytics.min.js > __tests__/prodsdk.js && jest --watchAll=false --maxWorkers=2 --coverage",
-    "start": "npm run build:integration:all && rollup --config rollup-configs/rollup.sdk.js --watch --environment DEV_SERVER",
-    "build:browser:dev": "rollup --config rollup-configs/rollup.sdk.js",
-    "build:browser": "rollup --config rollup-configs/rollup.sdk.js --environment VERSION:$npm_package_version,UGLIFY,PROD_DEBUG,STAGING:$npm_config_staging,ENV:prod",
-    "build:browser:modern": "BROWSERSLIST_ENV=modern rollup --config rollup-configs/rollup.sdk.js --environment VERSION:$npm_package_version,UGLIFY,PROD_DEBUG,STAGING:$npm_config_staging,ENV:prod",
-    "build:browser:debug": "rollup --config rollup-configs/rollup.sdk.js --environment VERSION:$npm_package_version,UGLIFY,PROD_DEBUG:inline,ENV:prod",
+    "test": "npm run build:browser && sed 's/var rudderanalytics/rudderanalytics/g' dist/legacy/rudder-analytics.min.js > __tests__/prodsdk.js && jest --watchAll=false --maxWorkers=50% --coverage",
+    "test:ci": "npm run build:browser && sed 's/var rudderanalytics/rudderanalytics/g' dist/legacy/rudder-analytics.min.js > __tests__/prodsdk.js && jest --watchAll=false --maxWorkers=2 --coverage",
+    "start": "npm run build:integration:all && rollup --config rollup-configs/rollup.sdk.browser.js --watch --environment DEV_SERVER",
+    "build:browser:dev": "rollup --config rollup-configs/rollup.sdk.browser.js",
+    "build:browser": "rollup --config rollup-configs/rollup.sdk.browser.js --environment VERSION:$npm_package_version,UGLIFY,PROD_DEBUG,STAGING:$npm_config_staging,ENV:prod",
+    "build:browser:modern": "BROWSERSLIST_ENV=modern rollup --config rollup-configs/rollup.sdk.browser.js --environment VERSION:$npm_package_version,UGLIFY,PROD_DEBUG,STAGING:$npm_config_staging,ENV:prod",
+    "build:browser:debug": "rollup --config rollup-configs/rollup.sdk.browser.js --environment VERSION:$npm_package_version,UGLIFY,PROD_DEBUG:inline,ENV:prod",
     "build:browser:bundle-size": "VISUALIZER=true npm run build:browser",
-    "build:node:dev": "rollup --config rollup-configs/rollup.sdk.js --environment VERSION:$npm_package_version,NODE_ENV,NPM",
-    "build:node": "rollup --config rollup-configs/rollup.sdk.js --environment VERSION:$npm_package_version,UGLIFY,ENV:prod,NODE_ENV,NPM",
+    "build:node:dev": "rollup --config rollup-configs/rollup.sdk.npm.js --environment VERSION:$npm_package_version,NODE_ENV,NPM",
+    "build:node": "rollup --config rollup-configs/rollup.sdk.npm.js --environment VERSION:$npm_package_version,UGLIFY,ENV:prod,NODE_ENV,NPM",
     "build:worker": "rollup --config rollup-configs/rollup.worker.config.js --environment VERSION:$npm_package_version,NPM",
     "build:npm": "npm run build:node && npm run build:worker",
     "build:integration:dev": "rollup --config rollup-configs/rollup.integrations.config.js --environment PROD_DEBUG,INTG_NAME:AdobeAnalytics",
