+++ rollup-configs/rollup.sdk.browser.js
@@ -1,12 +1,11 @@
 import { getOutputFilePath, getDefaultConfig } from './rollup.utilities';
-import { INTG_SUFFIX } from '../utils/constants';
 
 const variantSubfolder = process.env.BROWSERSLIST_ENV === 'modern' ? '/modern' : '/legacy';
 const sourceMapType =
   process.env.PROD_DEBUG === 'inline' ? 'inline' : process.env.PROD_DEBUG === 'true';
-const outDir = `dist${variantSubfolder}/integrations`;
-const distName = process.env.INTG_NAME;
-const modName = `${process.env.INTG_NAME}${INTG_SUFFIX}`;
+const outDir = `dist${variantSubfolder}`;
+const distName = 'rudder-analytics';
+const modName = 'rudderanalytics';
 
 const outFilePath = getOutputFilePath(outDir, distName);
 
@@ -19,8 +18,12 @@ const outputFiles = [
   },
 ];
 
-export default {
+const buildConfig = {
   ...getDefaultConfig(distName),
-  input: `integrations/${process.env.INTG_NAME}/index.js`,
+};
+
+export default {
+  ...buildConfig,
+  input: 'src/analytics.js',
   output: outputFiles,
 };
